Finished Survey (#32)
* Added global state for user.

* Updated login and startup script.

* Fixed issues with profile/login redirects.

* Added course lookup with endpoints. Updated login/general fixes.

* Fixed localhost dependencies in production mode.

* Fixed neo4jengine and edited survey for production.

* Updated get all courses endpoint to work from SQL.

* Updated survey validation.

* Cleaned up api and survey.

* Fixed issue with lookup.

* Added TODO.

* Fixed request issue.

* still doesnt work

* Fixed issue with data_util.

* Fixing yup for courses

* Fixing yup for courses

* Fixed autofill.

* Cleaned code, commented, fixed bug with semester input.

* Some fixes for production

* Safety checks for existence of server_info

* Removed unnecessary add buttons.

* Updated icon and index.html.

* Updated +/- buttons

* Fixed course name regex issue.

* Fixed issue with local storage.

* Updated schema.

* Updated readme.

* Major required validation - Yup

+++ server/setup.sh
@@ -1,5 +1,11 @@
 # Needs to be run with root permissions. Guaranteed to work only on an Amazon Linux Image.
 
+# require "server_info" file
+if [ ! -f /home/ec2-user/schedule-generator/server_info ]; then
+    echo "Need to add server_info to root directory of Git repo"
+    exit 1
+fi
+
 # install Apache2 HTTP server
 sudo yum update && sudo yum upgrade -y
 sudo yum install httpd tmux -y
@@ -51,7 +57,7 @@ cd /home/ec2-user/schedule-generator/server
 sudo cp ./httpd.conf /etc/httpd/conf/httpd.conf
 
 # setting up flask
-cd api
+cd /home/ec2-user/schedule-generator/server/api
 tmux new-session -d -s backend 'flask run --host 0.0.0.0 --port 5000'
 
 # run website
